template-skeleton: code-quality config + snippet library + doc-ref scrub (KIT-2)
PART A — safe doc-ref scrub (comment/docstring/string text ONLY, zero executable logic):
repointed/removed 9 dead refs to stripped docs (HARDENING-BACKLOG.md, docs/04-guardrails.md,
docs/13-mcp-integration.md) in agent/{guards,authz,a2a}.py, MASTERSCHEMA.md,
evals/datasets/{authz,injection_inert}.jsonl, evals/report/sample.json, requirements.txt
(kept the pip package list, dropped only the doc path). JSON stayed valid; imports unaffected.

PART B — code-quality config (pyproject.toml: ruff + black + mypy; .pre-commit-config.yaml; make lint):
- ruff (E,W,F,I,B,UP) tuned to RESPECT the skeleton's deliberate compact style — E701/E702/E703/E401
  ignored (one-line guards, semicolons, packed imports are voice, not bugs). black config present but
  NOT enforced (it would rewrite ~all files); pre-commit runs ruff only.
- fixed every genuine ruff finding: import-sort (I001) repo-wide, unused imports/vars, ambiguous names,
  zip(strict=False), useless f-strings, raise-from, lambda->def. 396 -> 0.
- mypy basic (ignore-missing-imports, no strict): fixed real type issues (implicit-Optional, None-guards,
  missing annotations, None-index); relaxed pure SDK-construction friction (agent.models LLM clients,
  evals.run_evals Braintrust Eval) per-module with comments. 43 -> 0.
- added requirements-dev.txt (ruff/black/mypy/pre-commit, pinned) so a bare clone can install the tools.

PART C — snippets/ paste-and-adapt templates (match real APIs, import real modules):
new_read_tool.py, new_action_tool.py, authz_rule.py, eval_cases.jsonl, telemetry_span.py, README.md.
BUILD-ORDER.md now points to them at SLOT 1 and SLOT 3.

Bare-clone gate stayed green throughout: update.py --check = 36 passed, 15 MANIFEST rows.
make lint green (ruff + mypy). No frame logic changed.

+++ tools/data_triage/profile.py
@@ -5,7 +5,11 @@ is inferred or fabricated (J-02). parquet/xlsx need optional readers (pyarrow/op
 are reported honestly, never faked.
 """
 from __future__ import annotations
-import csv, json, os, re
+
+import csv
+import json
+import os
+import re
 from collections import Counter
 
 SAMPLE_BYTES = 65536
@@ -20,7 +24,7 @@ _DATE = re.compile(r"^\d{4}-\d{2}-\d{2}([ T]\d{2}:\d{2}(:\d{2})?)?")
 _BOOL = {"true", "false", "yes", "no", "t", "f", "0", "1"}
 
 
-def _human(n: int) -> str:
+def _human(n: float) -> str:
     for u in ("B", "KB", "MB", "GB", "TB"):
         if n < 1024: return f"{n:.1f}{u}" if u != "B" else f"{n}B"
         n /= 1024
